+++ .gitignore
@@ -1,0 +1,2 @@
+src/jobs/*
+src/auth/credentials.json
