+++ src/structures/user.py
@@ -6,7 +6,7 @@ from src.auth.linkedin import initialize_linkedin_api
 
 class UserManager:
     def __init__(self):
-        self.file_handler = FileHandler('auth/credentials.json')
+        self.file_handler = FileHandler('../auth/credentials.json')
         self.encryptor = Encryptor("71X-ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789")  # Initialize the encryptor
         self.user_data = self._load_user_data()
