+++ src/utils/constants.py
@@ -1,0 +1,1 @@
+# Holds any constants
